backend/requirements.txt
fastapi
uvicorn
pydantic
pytest
pandas
numpy
python-dateutil
python-multipart
pyyaml
aiofiles
fastapi_utils
typing-inspect
requests